[#256] Refactor: Friend Command / Query 로직 분리
